requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.6
contourpy==1.0.5
cryptography==37.0.4
cycler==0.11.0
Flask==2.2.2
flatbuffers==2.0
flit_core==3.6.0
fonttools==4.25.0
frozendict==2.3.4
frozenlist==1.3.3
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.7.0
html5lib==1.1
idna==3.4
ipdb==0.13.9
itsdangerous==2.0.1
Jinja2==3.1.2
joblib==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
multitasking==0.0.9
munkres==1.1.4
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==22.0
pandas==1.4.2
Pillow==9.3.0
pip==22.3.1
ply==3.11
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.1.3
scipy==1.9.3
setuptools==65.6.3
sip==6.6.2
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
termcolor==2.1.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.1
yfinance==0.2.4
